requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
scikit-image>=0.21.0
opencv-python-headless>=4.8.0
einops>=0.6.0
timm>=0.9.0
gdown>=4.7.0

